Convenience methods for Element serialization, doc comments
