+++ src/lazy/binary/mod.rs
@@ -1,8 +1,9 @@
 mod encoded_value;
 pub mod immutable_buffer;
 pub mod lazy_reader;
-pub mod lazy_system_reader;
 pub mod raw;
+pub mod system;
 
+mod data_source;
 #[cfg(test)]
 mod test_utilities;
