+++ src/lazy/mod.rs
@@ -1,3 +1,6 @@
+//! Provides an ergonomic, lazy view of an Ion stream that permits random access within each
+//! top level value.
+
 pub mod binary;
 pub mod raw_stream_item;
 pub mod raw_value_ref;
